+++ server/src/routes/public.ts
@@ -1,5 +1,5 @@
 import express from "express";
-import { register } from "./controllers/register.js";
+import { register } from "../controllers/register.js";
 
 const router = express.Router();
