+++ server/src/server.ts
@@ -3,7 +3,7 @@ import express from "express";
 import { env } from "node:process";
 import { INTERNAL_SERVER_ERROR, NO_CONTENT } from "./http-status-code.js";
 import cors from "cors";
-import publicRouter from "./public-router.js";
+import publicRouter from "./routes/public.js";
 
 const CLIENT_ORIGIN = "https://¤CLIENT_DOMAIN_NAME¤";
 const PORT: number = parseInt(env.PORT ?? "3000", 10);
